refactor: ReservationWebRepository 서비스에 적용
